Update transaction to handle nonce from server
